backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.0
python-dotenv==1.0.1
gunicorn==23.0.0
flask-limiter>=3.5.0
flask-talisman>=1.1.0
email-validator>=2.0.0